Move resolvers/ folder
